Note to self, cast time_t to long int for printf... thanks Ankit for pointing this out.
git-svn-id: http://svn.inspircd.org/repository/trunk/inspircd@11012 e03df62e-2008-0410-955e-edbf42e46eb7

+++ src/modules/m_cban.cpp
@@ -47,7 +47,7 @@ public:
 
 	void DisplayExpiry()
 	{
-		ServerInstance->SNO->WriteToSnoMask('x',"Removing expired CBan %s (set by %s %ld seconds ago)", this->matchtext.c_str(), this->source, ServerInstance->Time() - this->set_time);
+		ServerInstance->SNO->WriteToSnoMask('x',"Removing expired CBan %s (set by %s %ld seconds ago)", this->matchtext.c_str(), this->source, (long int)(ServerInstance->Time() - this->set_time));
 	}
 
 	const char* Displayable()
